// frontend/src/api/models/MessageIn.ts
/**
 * Custom title
 * Here\'s a longer description of the custom **OpenAPI** schema
 *
 * OpenAPI spec version: 2.5.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { HttpFile } from '../http/http';

export class MessageIn {
    'conversationId': number;
    'content': string;
    'timestamp'?: Date;
    'userId': number;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "conversationId",
            "baseName": "conversation_id",
            "type": "number",
            "format": ""
        },
        {
            "name": "content",
            "baseName": "content",
            "type": "string",
            "format": ""
        },
        {
            "name": "timestamp",
            "baseName": "timestamp",
            "type": "Date",
            "format": "date-time"
        },
        {
            "name": "userId",
            "baseName": "user_id",
            "type": "number",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return MessageIn.attributeTypeMap;
    }

    public constructor() {
    }
}